requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-auth>=2.26.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
slack-sdk>=3.26.0
PyPDF2>=3.0.0
openai>=1.0.0
azure-core>=1.30.0
azure-search-documents>=11.4.0